fix: add missing code block in tutorial/orders
Moved code with orders tag from policy page to orders page
Fixes #2
